Fixed compiler warnings after uClibc code update

+++ src/libm/s_floor.c
@@ -58,7 +58,7 @@ double floor(double x)
 		    if(j0==20) i0+=1;
 		    else {
 			j = i1+(1<<(52-j0));
-			if(j<i1) i0 +=1 ; 	/* got a carry */
+			if(j<(u_int32_t)i1) i0 +=1 ; 	/* got a carry */
 			i1=j;
 		    }
 		}
